requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
dj-database-url==2.2.0
Django==5.0.7
django-crontab==0.7.1
django-environ==0.11.2
djangorestframework
gunicorn==22.0.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.1
psycopg2-binary==2.9.9
PySocks==1.7.1
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
